ErrandsTypes repository created
A repository has been made for event types (with or without repeatability, just hints or full-fledged tasks). Now I made it as a blank for a table, but I'm not sure that this part of the logic should be implemented by creating a separate table for types. Maybe it should be a new class, in which there is a method with the behaviour logic for push notifications written through ʼbools' and 'ifs'?

Signed-off-by: Tetiana <115114869+TetianaRaida@users.noreply.github.com>